// src/types/legacy.d.ts
export type Name = string;
export type Title = string;
export type Id = string;
export type Thumbnailurl = string;
// ...
export type Institution1 = string[];
/**
 * An enumeration.
 */
export type SoftwareAvailability = "open" | "partially open" | "closed";
export type Href = string;
export type Title1 = string;
export type Id2 = string;
export type Doi1 = string;
export type Doi = Hyperlink[];
export type Publications = Hyperlink[];
// ...
export type Views = DatasetView[];

export interface DatasetManifest {
  name: Name;
  metadata: DatasetMetadata;
  sources: Sources;
  views: Views;
}
/**
 * Metadata for a bioimaging dataset.
 */
export interface DatasetMetadata {
  title: Title;
  id: Id;
  thumbnailURL?: Thumbnailurl;
  imaging: FIBSEMImagingMetadata;
  sample: SampleMetadata;
  institution: Institution1;
  softwareAvailability: SoftwareAvailability;
  DOI: Doi;
  publications: Publications;
}
/**
 * Metadata describing the FIB-SEM imaging process.
 */
export interface FIBSEMImagingMetadata {
  id: Id1;
  institution: Institution;
  gridSpacing: UnitfulVector;
  dimensions: UnitfulVector;
  startDate: Startdate;
  duration: Duration;
  biasVoltage: Biasvoltage;
  scanRate: Scanrate;
  current: Current;
  primaryEnergy: Primaryenergy;
}
export interface UnitfulVector {
  unit: Unit;
  values: Values;
}
export interface Values {
  [k: string]: number;
}
/**
 * Metadata describing the sample and sample preparation.
 */
export interface SampleMetadata {
  description: Description;
  protocol: Protocol;
  contributions: Contributions;
  organism: Organism;
  type: Type;
  subtype: Subtype;
  treatment: Treatment;
}
export interface Hyperlink {
  href: Href;
  title: Title1;
}
export interface DOI {
  id: Id2;
  DOI: Doi1;
}
export interface Sources {
  [k: string]: VolumeSource;
}
export interface VolumeSource {
  name: Name1;
  description: Description1;
  url: Url;
  format: ArrayContainerTypeEnum;
  transform: SpatialTransform;
  sampleType: SampleTypeEnum;
  contentType: ContentTypeEnum;
  displaySettings: DisplaySettings;
  subsources: Subsources;
}
// ...
export interface DatasetView {
  name: Name3;
  description: Description3;
  sources: string[];
  position?: Position;
  scale?: Scale1;
  orientation?: Orientation;
}